Updating NpgsqlCommand test with snapshots and removing DBM file
